Move Authenticators fixture under Authenticator folder
